Re-creating the rebase PR again form PR 869 , as by mistake we did
squash and merge . so trying to fix in another branch and than force
push to amd-npi

This PR merges the latest changes from
auto-merge-rocm-systems-develop-to-amd-npi-826 into
amd-npi-compute-external-merge-new-826

+++ projects/amdsmi/goamdsmi_shim/smiwrapper/amdsmi_go_shim.c
@@ -462,32 +462,32 @@ uint32_t goamdsmi_cpu_prochot_status_get(uint32_t socket_index) {
 
 uint32_t goamdsmi_cpu_socket_power_get(uint32_t socket_index) {
   bool readSuccess = false;
-  double socket_power_watts = 0;
+  uint32_t socket_power_watts = 0;
   if ((AMDSMI_STATUS_SUCCESS ==
        amdsmi_get_cpu_socket_power(amdsmi_processor_handle_all_cpu_across_socket[socket_index],
                                    &socket_power_watts)))
     readSuccess = true;
   if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {
-    printf("AMDSMI, %s for Socket:%d, CpuSocketPowerWatt:%.3f\n",
-           readSuccess ? "Success" : "Failed", socket_index, socket_power_watts);
+    printf("AMDSMI, %s for Socket:%d, CpuSocketPowerWatt:%u\n", readSuccess ? "Success" : "Failed",
+           socket_index, socket_power_watts);
   }
 
-  return readSuccess ? (uint32_t)(socket_power_watts * 1000) : GOAMDSMI_UINT32_MAX;
+  return readSuccess ? socket_power_watts * 1000 : GOAMDSMI_UINT32_MAX;
 }
 
 uint32_t goamdsmi_cpu_socket_power_cap_get(uint32_t socket_index) {
   bool readSuccess = false;
-  double socket_power_cap_watts = 0;
+  uint32_t socket_power_cap_watts = 0;
   if ((AMDSMI_STATUS_SUCCESS ==
        amdsmi_get_cpu_socket_power_cap(amdsmi_processor_handle_all_cpu_across_socket[socket_index],
                                        &socket_power_cap_watts)))
     readSuccess = true;
   if (enable_debug_level(GOAMDSMI_DEBUG_LEVEL_1)) {
-    printf("AMDSMI, %s for Socket:%d, CpuSocketPowerCapWatt:%.3f\n",
+    printf("AMDSMI, %s for Socket:%d, CpuSocketPowerCapWatt:%u\n",
            readSuccess ? "Success" : "Failed", socket_index, socket_power_cap_watts);
   }
 
-  return readSuccess ? (uint32_t)(socket_power_cap_watts * 1000) : GOAMDSMI_UINT32_MAX;
+  return readSuccess ? socket_power_cap_watts * 1000 : GOAMDSMI_UINT32_MAX;
 }
 
 uint32_t goamdsmi_cpu_core_boostlimit_get(uint32_t thread_index) {
